Add sortedCharacterList
For some reason data that uses .ToListAsync() (used with async function) can't be recognize as the same object type used for the output when OrderBy() is used therefore, we have to pass the data to another variable and use .ToList().

This way, we can use await in async and sort the data before returning the data to the client.